made changes such that logging file was created in src and functions were called from it in scripts
