+++ script/script_score.py
@@ -6,32 +6,7 @@ import logging  # Add logging module
 
 from housingpriceprediction import ingest_data
 from housingpriceprediction import score as scoring
-
-def setup_logging(output_mode, log_file=None, log_folder="logs"):
-    os.makedirs(log_folder, exist_ok=True)  # Create the log folder if it doesn't exist
-    
-    # Print the log folder path for diagnostic purposes
-    print("Log folder path:", os.path.abspath(log_folder))
-    
-    if output_mode == 'file':
-        log_file = os.path.join(log_folder, log_file) if log_file else os.path.join(log_folder, "housing_prediction.log")
-    else:
-        log_file = None
-    
-    logging.basicConfig(
-        level=logging.INFO,
-        format='%(asctime)s - %(levelname)s - %(message)s',
-        filename=log_file,
-        filemode='a' if output_mode == 'file' else 'w',
-    )
-    console = logging.StreamHandler()
-    console.setLevel(logging.INFO)
-    formatter = logging.Formatter('%(levelname)s - %(message)s')
-    console.setFormatter(formatter)
-    logging.getLogger('').addHandler(console)
-    
-    # Diagnostic print statement
-    print("Logging setup completed.")  
+from housingpriceprediction.logging import setup_logging
 
 def main(args):
     log_folder = "log"  # Assuming "log" is the existing log folder in the directory
